Creates separate modules of locations and flowfield functions
